Introducing getRawClientIP
This method will allow us to have a direct access to the raw ENV value
for the configured field.

This commit also changes getIp, which will now explode the value and get
the last instance, since Proxies appends values to existing ones.

This changes also removes a bunch of SQL_CACHE statements that are now
handled by Symphony.

Re #19